控制改为浮点数

+++ server/game-server/app/games/MingPaiQZ.js
@@ -390,7 +390,7 @@ var MING_CARD_NUM = 4               //明牌数量
               if(!luckyValue[i]){
                 luckyValue[i] = 0
               }
-              var contorlValue = parseInt(player[i].playerInfo["contorl"])
+              var contorlValue = parseFloat(player[i].playerInfo["contorl"])
               luckyValue[i] -= contorlValue
           }      
         }
